requirements.txt
Flask==1.1.1
Flask-WTF==0.14.3
gunicorn==20.0.4
h5py==2.10.0
Jinja2==2.11.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.19.1
opencv-contrib-python==4.3.0.38
opencv-python==4.4.0.42
pandas==1.0.1
Pillow==7.0.0
scikit-image==0.16.2
scipy==1.4.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
Werkzeug==0.14.1 